Address third-round PR #882 review feedback
- Move the DomainMapping / DomainToEntityIdMapping / ResolvedDomain type
  definitions and the validation functions to a new
  src/schema/entity-domains.ts library; data/entities/domains/
  entity-domains.ts keeps only the global mapping singleton
  (updateDomainMapping, domainMappingForDomain, entitiesForDomain).
  This puts the interface definitions outside the data/** sort-keys
  rule, so the eslint-disable directives are gone and
  eslint-plugin-sort-keys-custom-order is reverted to ^2.2.1
  (package.json and pnpm-lock.yaml are back to identical with beta).
- readDomainMappingFile is a pure read again; the module-level mapping
  is now refreshed in writeDomainMappingFile, i.e. whenever
  entity-domains.json is overwritten.
