+++ package.json
@@ -82,7 +82,7 @@
 		"eslint-config-love": "^145.0.0",
 		"eslint-config-prettier": "^10.1.8",
 		"eslint-plugin-simple-import-sort": "^12.1.1",
-		"eslint-plugin-sort-keys-custom-order": "^3.0.1",
+		"eslint-plugin-sort-keys-custom-order": "^2.2.1",
 		"eslint-plugin-svelte": "^3.19.0",
 		"globals": "^17.4.0",
 		"harper.js": "^0.61.0",
